feat: split consolidations service
